Extract check_cache into own file with optional args
